requirements.txt
# Core dependencies for Zola-mac content automation

requests>=2.31.0
yt-dlp>=2023.0.0
pytube>=15.0.0
python-slugify>=8.0.0
python-dotenv>=1.0.0
readability-lxml>=0.8.1
toml>=0.10.0
Pillow>=10.0.0

# Text-to-Speech (Edge-TTS)
edge-tts==7.0.0
langdetect>=1.0.9
tqdm>=4.67.1
pydub>=0.25.1

# Text-to-Speech (Gemini TTS - Primary)
google-genai>=1.49.0

# Optional legacy fallback
# pyttsx3>=2.90  # Commented out — replaced by neural Edge-TTS

# FFmpeg required (not pip-installable)
# macOS: brew install ffmpeg
# Linux: sudo apt install ffmpeg
